refactor : 설문제작 페이지 리펙토링
