+++ src/components/createForm/createFormMain/components/questionTypesForm/LongAndShortDescriptiveForm.jsx
@@ -1,25 +1,22 @@
-import React, { useEffect, useState } from "react";
-import { useDispatch, batch } from "react-redux";
+import React, { useEffect } from "react";
+import { useDispatch, useSelector, batch } from "react-redux";
 import styled from "styled-components";
-import RoundButtonLarge from "../../../../common/buttons/roundButtons/RoundButtonLarge";
-import { useSelector } from "react-redux";
+
 import {
   selectedFormType,
   fillOutQuestion,
 } from "../../../../../redux/modules/createFormSlice";
+import RoundButtonLarge from "../../../../common/buttons/roundButtons/RoundButtonLarge";
 import fonts from "../../../../../styles/fonts";
 
 const LongAndShortDescriptiveForm = () => {
   const dispatch = useDispatch();
-  // 현재 타입
   const questionType = useSelector(
     (state) => state.createForm.selectedFormType
   );
-  // 현재 페이지 넘버
   const currentPageNum = useSelector(
     (state) => state.createForm.currentPageNum
   );
-  // 타이틀이 있을때면 기존에 값이 있다 판단, 삭제 하면 안되기 때문에 체크 포인트로 사용
   const questionTitle = useSelector(
     (state) =>
       state.createForm.formList?.questionList[currentPageNum - 2][
@@ -27,7 +24,6 @@ const LongAndShortDescriptiveForm = () => {
       ]
   );
 
-  // 처음 시작할 때 세팅 값
   useEffect(() => {
     currentPageNum > 1 &&
       !questionTitle &&
@@ -41,7 +37,6 @@ const LongAndShortDescriptiveForm = () => {
       );
   }, []);
 
-  // 내부에서 형식을 바꿨을 때
   const typeClickHandler = (type) => {
     batch(() => {
       dispatch(
@@ -83,11 +78,12 @@ const LongAndShortDescriptiveForm = () => {
 };
 
 const Container = styled.div`
-  width: 100%;
   display: flex;
   flex-direction: column;
   align-items: center;
+
   margin-top: 11.6rem;
+  width: 100%;
   p {
     ${fonts.Body1}
     font-weight: 500;
