+++ src/components/createForm/createFormMain/components/questionTypesForm/RankForm.jsx
@@ -1,23 +1,20 @@
 import React, { useEffect, useState } from "react";
-import styled from "styled-components";
-import fonts from "../../../../../styles/fonts";
 import { useSelector, useDispatch } from "react-redux";
+import styled from "styled-components";
+
 import { fillOutQuestion } from "../../../../../redux/modules/createFormSlice";
+import fonts from "../../../../../styles/fonts";
 
 const RankForm = () => {
   const dispatch = useDispatch();
-  // 입력값
   const [questionInput, setQuestion] = useState("");
-  // 현재 페이지 넘버
   const currentPageNum = useSelector(
     (state) => state.createForm?.currentPageNum
   );
-  // 답으로 체크할 항목
   const answerList = useSelector(
     (state) =>
       state.createForm.formList?.questionList[currentPageNum - 2]["answerList"]
   );
-  // 타이틀이 있을 때면 기존에 값이 있다 판단, 삭제 하면 안되기 때문에 체크 포인트로 사용
   const questionTitle = useSelector(
     (state) =>
       state.createForm.formList?.questionList[currentPageNum - 2][
@@ -25,7 +22,6 @@ const RankForm = () => {
       ]
   );
 
-  // 처음 시작할 때 세팅 값
   useEffect(() => {
     currentPageNum > 1 &&
       !questionTitle &&
@@ -39,27 +35,21 @@ const RankForm = () => {
       );
   }, []);
 
-  // 페이지가 넘어가면 input값 초기화
   useEffect(() => {
     setQuestion("");
   }, [currentPageNum]);
 
-  // answer 값 수정
   const InputChangeHandler = (event) => {
     setQuestion(event.target.value);
   };
 
-  // 엔터 버튼 눌렀을 때
   const onKeyUp = (event) => {
     if (event.target.value.length !== 0 && event.key === "Enter") {
       submitTagItem();
     }
   };
   const submitTagItem = () => {
-    dispatch(
-      // fillOutQuestionAnswer
-      fillOutQuestion({ answerList: [...answerList, questionInput] })
-    );
+    dispatch(fillOutQuestion({ answerList: [...answerList, questionInput] }));
     setQuestion("");
   };
 
@@ -93,10 +83,11 @@ const RankForm = () => {
 };
 
 const Container = styled.div`
-  width: 100%;
-  margin-top: 5rem;
   display: flex;
   justify-content: center;
+
+  margin-top: 5rem;
+  width: 100%;
 `;
 
 const ChoiceContainer = styled.div`
@@ -114,25 +105,28 @@ const QuestionInput = styled.input`
   font-weight: 500;
   font-size: 1.6rem;
   line-height: 1.9rem;
-  border: none;
+
   width: 100%;
   padding: 0.8rem 0;
+
+  border: none;
   border-bottom: ${({ theme }) => `0.2rem solid ${theme.gray3}`};
 `;
 
 const Question = styled.div`
   display: flex;
   align-items: center;
+
   ${fonts.Body1}
   font-weight: 500;
   font-size: 1.6rem;
   line-height: 1.9rem;
-  border: none;
-  width: 100%;
-  margin-top: 2.2rem;
+
   padding: 0.2rem 0;
   margin-top: 1rem;
   width: 100%;
+
+  border: none;
   border-bottom: ${({ theme }) => `0.2rem solid ${theme.gray3}`};
   div {
     flex: 1;
@@ -140,12 +134,11 @@ const Question = styled.div`
     font-weight: 500;
     font-size: 1.6rem;
     line-height: 1.9rem;
-
     span {
+      margin-right: 1rem;
       font-weight: 800;
       font-size: 2rem;
       line-height: 1.9rem;
-      margin-right: 1rem;
     }
   }
   img {
