+++ src/components/createForm/createFormMain/components/questionTypesForm/ScoreAndStarForm.jsx
@@ -1,24 +1,22 @@
-import React, { useEffect, useState } from "react";
-import styled from "styled-components";
-import RoundButtonLarge from "../../../../common/buttons/roundButtons/RoundButtonLarge";
+import React, { useEffect } from "react";
 import { useSelector, useDispatch, batch } from "react-redux";
+import styled from "styled-components";
+
 import {
   selectedFormType,
   fillOutQuestion,
 } from "../../../../../redux/modules/createFormSlice";
 import fonts from "../../../../../styles/fonts";
+import RoundButtonLarge from "../../../../common/buttons/roundButtons/RoundButtonLarge";
 
 const ScoreAndStarForm = () => {
   const dispatch = useDispatch();
-  // 현재 타입
   const questionType = useSelector(
     (state) => state.createForm.selectedFormType
   );
-  // 현재 페이지 넘버
   const currentPageNum = useSelector(
     (state) => state.createForm.currentPageNum
   );
-  // 타이틀이 있을때면 기존에 값이 있다 판단, 삭제 하면 안되기 때문에 체크 포인트로 사용
   const questionTitle = useSelector(
     (state) =>
       state.createForm.formList?.questionList[currentPageNum - 2][
@@ -26,9 +24,6 @@ const ScoreAndStarForm = () => {
       ]
   );
 
-  console.log(questionType);
-
-  // 처음에 시작할 때 셋팅 값
   useEffect(() => {
     currentPageNum > 1 &&
       !questionTitle &&
@@ -42,7 +37,6 @@ const ScoreAndStarForm = () => {
       );
   }, []);
 
-  // 내부에서 형식을 바꿨을 때
   const typeClickHandler = (type) => {
     batch(() => {
       dispatch(
@@ -79,11 +73,12 @@ const ScoreAndStarForm = () => {
 };
 
 const Container = styled.div`
-  width: 100%;
   display: flex;
   flex-direction: column;
   align-items: center;
+
   margin-top: 11.6rem;
+  width: 100%;
   p {
     ${fonts.Body1}
     font-weight: 500;
