+++ src/components/createForm/createFormPreview/components/CreateFormMobile.jsx
@@ -10,13 +10,15 @@ const CreateFormMobile = () => {
   );
 };
 const Container = styled.div`
-  width: 35.6rem;
-  height: 77.3rem;
-  box-shadow: 0px 0px 7px 3px rgba(0, 0, 0, 0.25);
-  border-radius: 44px;
   display: flex;
   justify-content: center;
   align-items: center;
+
+  width: 35.6rem;
+  height: 77.3rem;
+
+  box-shadow: 0px 0px 7px 3px rgba(0, 0, 0, 0.25);
+  border-radius: 44px;
 `;
 
 export default CreateFormMobile;
