+++ src/components/createForm/createFormPreview/components/typeOfAnswer/CoverAnswer.jsx
@@ -1,7 +1,8 @@
 import React from "react";
+import { useSelector } from "react-redux";
 import styled from "styled-components";
+
 import fonts from "../../../../../styles/fonts";
-import { useDispatch, useSelector } from "react-redux";
 import RoundButtonLarge from "../../../../common/buttons/roundButtons/RoundButtonLarge";
 
 const CoverAnswer = () => {
@@ -31,34 +32,35 @@ const CoverAnswer = () => {
 };
 
 const Container = styled.div`
-  width: 100%;
-  height: 100%;
   position: relative;
   display: flex;
   flex-direction: column;
+
+  width: 100%;
+  height: 100%;
 `;
 
 const Header = styled.div`
-  text-align: center;
   margin-top: 6.1rem;
+  text-align: center;
   h1 {
+    margin: 0;
     ${fonts.Body1}
     font-weight: 700;
     font-size: 2.4rem;
     line-height: 2.9rem;
-    margin: 0;
   }
 `;
 
 const Main = styled.div`
   flex: 1;
-  text-align: center;
-  margin-top: 4.6rem;
   display: flex;
   flex-direction: column;
   justify-content: space-between;
-  height: 20rem;
 
+  margin-top: 4.6rem;
+  height: 20rem;
+  text-align: center;
   div {
     height: 13rem;
     ${fonts.Body3}
