+++ src/components/createForm/createFormPreview/components/typeOfAnswer/ShortDescriptiveAnswer.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import styled from "styled-components";
+
 import fonts from "../../../../../styles/fonts";
 
 const ShortDescriptiveAnswer = () => {
@@ -11,22 +12,23 @@ const ShortDescriptiveAnswer = () => {
 };
 
 const Container = styled.div`
-  width: 100%;
   display: flex;
   justify-content: center;
   align-items: center;
+
   margin: 7rem;
+  width: 100%;
   input {
-    text-align: center;
     padding: 1.1rem;
-    width: 27.3rem;
     margin-top: 1.5rem;
+    width: 27.3rem;
 
     ${fonts.Body1}
     font-weight: 600;
     font-size: 1.4rem;
     line-height: 1.7rem;
 
+    text-align: center;
     border: none;
     border-bottom: 0.3rem solid ${({ theme }) => theme.mainColor};
     &::placeholder {
