+++ src/components/createForm/createFormPreview/components/typeOfAnswer/StarAnswer.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import styled from "styled-components";
+
 import fonts from "../../../../../styles/fonts";
 
 const StarAnswer = () => {
@@ -47,15 +48,16 @@ const StarAnswer = () => {
   );
 };
 const Container = styled.div`
-  width: 100%;
   display: flex;
-  margin-top: 11.7rem;
   justify-content: center;
   align-items: center;
   flex-direction: column;
+
+  margin-top: 11.7rem;
+  width: 100%;
   p {
-    ${fonts.Body3}
     margin-top: 1.9rem;
+    ${fonts.Body3}
     font-weight: 400;
     font-size: 1.2rem;
     line-height: 1.4rem;
@@ -65,22 +67,23 @@ const Container = styled.div`
 const ScoreButtonContainer = styled.div`
   display: flex;
   div {
-    margin: 0 0.5rem 0 0.5rem;
     display: flex;
     flex-direction: column;
     align-items: center;
+
+    margin: 0 0.5rem;
     cursor: pointer;
     img {
       width: 4.8rem;
     }
     h3 {
+      margin: 1rem 0 0 0;
+      padding: 0;
+
       ${fonts.Body1}
       font-weight: 400;
       font-size: 1.6rem;
       line-height: 1.9rem;
-      margin-top: 1rem;
-      margin-bottom: 0;
-      padding: 0;
     }
   }
 `;
